+++ sama/model.py
@@ -40,9 +40,17 @@ class Sale(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     date: str
     product_id: int
+    product_name: str
     sell_price: float
     cost: float
     qty: int
+
+    def __str__(self):
+        """
+        Representation modification.
+        :return:
+        """
+        return f"{self.id} ({self.date})"
 
 
 def _verify_db() -> None:
@@ -62,17 +70,6 @@ def add_product(product: Product) -> str:
         return f"Error: {pformat(ex)}"
 
 
-def add_sale(sale: Sale) -> str:
-    """Add sale to db."""
-    try:
-        with Session(engine) as session:
-            session.add(sale)
-            session.commit()
-        return "Product has been saved!"
-    except Exception as ex:
-        return f"Error: {pformat(ex)}"
-
-
 def get_product(id_: int) -> Optional[Product]:
     """Return a product by id."""
     try:
@@ -84,43 +81,57 @@ def get_product(id_: int) -> Optional[Product]:
         return None
 
 
-def update_product(
-    id_: int,
-    qty: Optional[int] = None,
-    name: Optional[str] = None,
-    cost: Optional[float] = None,
-    price: Optional[float] = None,
-    tag: Optional[str] = None,
-) -> Optional[Product]:
-    """Update a product property."""  # TODO
+def get_all_products() -> Optional[List[Product]]:
+    """Return all products from db."""
     try:
         with Session(engine) as session:
-            statement = select(Product).where(Product.id == id_)
-            product = session.exec(statement).one()
-            if qty:
-                product.qty = qty
-            if name:
-                product.name = name
-            if cost:
-                product.cost = cost
-            if price:
-                product.price = price
-            if tag:
-                product.tag = tag
-            session.add(product)
-            session.commit()
+            statement = select(Product).order_by(Product.name)
+            results = session.exec(statement)
+            return results.all()
     except Exception as ex:
         print(f"Error: {pformat(ex)}")
         return None
 
 
-def get_all_products() -> Optional[List[Product]]:
-    """Return all products from db."""
+def update_product(product: Product) -> None:
+    """Update a product property."""
     try:
         with Session(engine) as session:
-            statement = select(Product)
-            results = session.exec(statement)
-            return results.all()
+            session.add(product)
+            session.commit()
+    except Exception as ex:
+        print(f"Error: {pformat(ex)}")
+
+
+def delete_product(product: Product) -> None:
+    """Delete product from db."""
+    try:
+        with Session(engine) as session:
+            statement = select(Product).where(Product.id == product.id)
+            result = session.exec(statement).one()
+            session.delete(result)
+            session.commit()
+    except Exception as ex:
+        print(f"Error: {pformat(ex)}")
+
+
+def add_sale(sale: Sale) -> str:
+    """Add sale to db."""
+    try:
+        with Session(engine) as session:
+            session.add(sale)
+            session.commit()
+        return "Product has been saved!"
+    except Exception as ex:
+        return f"Error: {pformat(ex)}"
+
+
+def get_sale(id_: int) -> Optional[Sale]:
+    """Return a sale record by id."""
+    try:
+        with Session(engine) as session:
+            statement = select(Sale).where(Sale.id == id_)
+            return session.exec(statement).one()
     except Exception as ex:
         print(f"Error: {pformat(ex)}")
         return None
@@ -130,12 +141,34 @@ def get_all_sales() -> Optional[List[Sale]]:
     """Return all products from db."""
     try:
         with Session(engine) as session:
-            statement = select(Sale)
+            statement = select(Sale).order_by(Sale.date)
             results = session.exec(statement)
             return results.all()
     except Exception as ex:
         print(f"Error: {pformat(ex)}")
         return None
+
+
+def update_sale(sale: Sale) -> None:
+    """Update sale to db."""
+    try:
+        with Session(engine) as session:
+            session.add(sale)
+            session.commit()
+    except Exception as ex:
+        print(f"Error: {pformat(ex)}")
+
+
+def delete_sale(sale: Sale) -> None:
+    """Delete sale to db."""
+    try:
+        with Session(engine) as session:
+            statement = select(Sale).where(Sale.id == sale.id)
+            result = session.exec(statement).one()
+            session.delete(result)
+            session.commit()
+    except Exception as ex:
+        print(f"Error: {pformat(ex)}")
 
 
 if __name__ == "__main__":