+++ CHANGELOG.md
@@ -5,9 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.2] - 2023-07-05
+
+### Added
+- Grouped commands for inventory and sales management, `stock` and `store`.
+- Update and Delete functions for product and sale records.
+
+
 ## [0.0.1] - 2023-06-30
 
 ### Added
 - Sama: A simple command line tool to manage sales and inventory for small retail business.
     - Add products to inventory.
-    - Register sales. 
+    - Register sale records.